move custom client fixture up a level
